// src/models/enums/modelName.ts
export enum ModelName {
  HEALTH_GUIDE = 'health_guide',
  HEALTH_CHOICE = 'health_choice',
  HEALTH_GOAL = 'health_goal',
  HEALTH_SYMPTOM = 'health_symptom',
  HEALTH_CONDITION = 'health_condition',
  HEALTH_PHYSICAL_ACTIVITY_LEVEL = 'health_physical_activity_level',
  HEALTH_DIET = 'health_diet',
  PERSONAL_HEALTH_QUERY = 'personal_health_query',
  FITNESS_EXERCISE_GROUP = 'fitness_exercise_group',
  FITNESS_PLAN = 'fitness_plan',
  FITNESS_EXERCISE = 'fitness_exercise',
  NUTRITION_GUIDE = 'nutrition_guide',
  PERIOD_DATA = 'period_data',
  USER_TIMEZONE = 'user_timezone',
}